version 1.4.0

+++ src/Manage.php
@@ -141,7 +141,12 @@ class Manage
         }
         $newTask = $this->run_task;
         foreach($killList as $k=>$v){
-            $crontabRule = $this->getCrontabRule($this->keyToClass($k));
+            $keyToClass = $this->keyToClass($k);
+            if(!$keyToClass){
+                echo get_color_text(31,'class '.$k.' is not exists').PHP_EOL;
+                continue;
+            }
+            $crontabRule = $this->getCrontabRule($keyToClass);
             CrontabManage::getInstance()->delete($crontabRule);
             unset($newTask[$k]);
         }
@@ -231,7 +236,16 @@ class Manage
             //记录日志
             return get_color_text(31, 'warning：Task' . $taskInfo['class'] . 'setCrontab returns abnormal format!' . PHP_EOL);
         }
-        $crontabRule[5] =             $crontabRule[5] = 'php ' . SNAN_BIN_PATH . ' run ' . uncamelize($taskInfo['class']).' > '.SNANTASK_ROOT.'/snantask_error.log';
+        $mode = strtolower(env_val('task_config.mode','default'));
+        $nameSpace = str_replace('\\','\\\\',env_val('task_config.task_namespace','\\SnanWord\\task\\'));
+        switch ($mode){
+            case 'think':
+                $crontabRule[5] = $crontabRule[5] = 'php ' . SNANTASK_ROOT . '/think snantask ' .$nameSpace. $taskInfo['class'].' > '.SNANTASK_ROOT.'/snantask_error.log';
+                break;
+            default:
+                $crontabRule[5] =             $crontabRule[5] = 'php ' . SNAN_BIN_PATH . ' run ' . uncamelize($taskInfo['class']).' > '.SNANTASK_ROOT.'/snantask_error.log';
+                break;
+        }
         return $crontabRule;
     }
 }
